i-nergy-load-forecasting-nbeats/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
boto3==1.21.42
botocore==1.24.42
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
colorama==0.4.4
convertdate==2.4.0
cycler==0.10.0
Cython==0.29.28
databricks-cli==0.16.6
decorator==5.1.1
distlib==0.3.2
docker==5.0.3
entrypoints==0.4
executing==0.8.3
filelock==3.0.12
Flask==2.1.1
fonttools==4.32.0
frozenlist==1.3.0
fsspec==2022.3.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.5
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.44.0
hijri-converter==2.2.3
holidays==0.13
idna==3.3
importlib-metadata==4.11.3
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
lightgbm==3.3.2
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mlflow==1.25.1
multidict==6.0.2
nfoursid==1.0.0
numpy==1.21.2
oauthlib==3.2.0
packaging==21.3
pandas==1.3.3
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.3.0
plotly==5.3.1
pmdarima==1.8.5
pretty-errors==1.2.25
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.1
prompt-toolkit==3.0.29
protobuf==3.20.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyDeprecate==0.3.2
Pygments==2.11.2
PyJWT==2.3.0
PyMeeus==0.5.11
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-lightning==1.6.1
pytz==2021.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.7.2
waitress==2.1.1
wcwidth==0.2.5
websocket-client==1.3.2
Werkzeug==2.1.1
xarray==2022.3.0
yarl==1.7.2
zipp==3.8.0
u8darts==0.18.0
